Introducing more button functionality
Introduced more button functionality. For example, allowing the publisher to set a registration end date. This provides additional functionality for the publishers as a whole.

Also cleaned up redirecting 404 bugs, as some of the past webinars are hardcoded.

Hiding content from mobile view for the card-variant-event-information to improve user experience